Afegit readme al 4.1
