--- NODE/SPRINT4/SPRINT 4.1/.env-template
@@ -1,15 +1,0 @@
-# database connection
-DATABASE_HOST=localhost
-DATABASE_PORT=3306
-DATABASE_USER=root
-DATABASE_PASSWORD=1234
-DATABASE_NAME=test_db
-
-# run options (mysql, mongodb, redis)
-DATABASE=mysql
-
-# credentials
-GOOGLE_ACCOUNT = youremail@gmail.com
-GOOGLE_APP_PASSWORD = yourpassword
-GOOGLE_API_KEY = yourkey
-GOOGLE_SECRET_KEY = yoursecretkey
